service/bot_api/pyproject.toml
[tool.poetry]
name = "bot_api"
version = "0.1.0"
description = ""
authors = ["Peter Ibragimov"]

[tool.poetry.dependencies]
python = "^3.9"
ujson = "^4.3.0"
aws-lambda-powertools = "^1.28.0"
python-telegram-bot = {extras = ["all"], version = "^20.0"}
pydantic = {extras = ["dotenv"], version = "^1.10.4"}
uvloop = "^0.17.0"
nest-asyncio = "^1.5.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
